🌱 Split kubectlProxy.core.test.ts into 2 focused test files
Split 1975-line kubectlProxy.core.test.ts into:
- kubectlProxy.core.connection.test.ts (~838 lines): connection lifecycle, timeout, cooldown, exec, queue, close
- kubectlProxy.core.resources.test.ts (~1338 lines): getNodes, getPodMetrics, getServices, getPVCs, getClusterHealth, getPodIssues, getEvents, getDeployments, etc.

All 51 tests pass. No new lint warnings. Part of #10439.
